+++ mapillary_tools/geotag/blackvue_parser.py
@@ -7,7 +7,7 @@ import typing as T
 import pynmea2
 
 from .. import geo
-from . import simple_mp4_parser as sparser
+from ..mp4 import simple_mp4_parser as sparser
 
 
 LOG = logging.getLogger(__name__)
