+++ mapillary_tools/geotag/gpmf_parser.py
@@ -5,7 +5,7 @@ import typing as T
 import construct as C
 
 from .. import geo
-from . import mp4_sample_parser as sample_parser
+from ..mp4 import mp4_sample_parser as sample_parser
 
 """
 Parsing GPS from GPMF data format stored in GoPros. See the GPMF spec: https://github.com/gopro/gpmf-parser
