+++ mapillary_tools/mp4/construct_mp4_parser.py
