+++ mapillary_tools/mp4/simple_mp4_parser.py
